another quick update, QOL and restructuring for the first product to go out officially -datafeed automapping2.0
