+++ src/App.js
@@ -1,44 +1,10 @@
 import React from "react";
-import { HashRouter as Router, Route, Routes, Link } from "react-router-dom";
+import { HashRouter as Router, Route, Routes } from "react-router-dom";
 import "bootstrap/dist/css/bootstrap.min.css";
 import Automapper from "./pages/Automapper.js";
 import Home from "./pages/home.js";
 import FishUSA from "./pages/FishUSA.js";
-
-const Navbar = () => (
-	<nav className="navbar navbar-expand-lg navbar navbar-dark bg-primary fixed-top">
-		<div className="container ">
-			<Link className="navbar-brand" to="/">
-				AvantLink Support App
-			</Link>
-			<button
-				className="navbar-toggler"
-				type="button"
-				data-bs-toggle="collapse"
-				data-bs-target="#navbarNav"
-				aria-controls="navbarNav"
-				aria-expanded="false"
-				aria-label="Toggle navigation"
-			>
-				<span className="navbar-toggler-icon"></span>
-			</button>
-			<div className="collapse navbar-collapse" id="navbarNav">
-				<ul className="navbar-nav">
-					<li className="nav-item">
-						<Link className="nav-link" to="/">
-							Home
-						</Link>
-					</li>
-					<li className="nav-item">
-						<Link className="nav-link" to="/automapper">
-							Automapper
-						</Link>
-					</li>
-				</ul>
-			</div>
-		</div>
-	</nav>
-);
+import Navbar from "./components/NavBar.js";
 
 function App() {
 	return (
