+++ src/components/NavBar.js
@@ -1,0 +1,44 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+const Navbar = () => (
+	<nav className="navbar navbar-expand-lg navbar navbar-dark bg-primary fixed-top">
+		<div className="container">
+			<Link className="navbar-brand" to="/">
+				Chetti.Tools Support App
+			</Link>
+			<button
+				className="navbar-toggler"
+				type="button"
+				data-bs-toggle="collapse"
+				data-bs-target="#navbarNav"
+				aria-controls="navbarNav"
+				aria-expanded="false"
+				aria-label="Toggle navigation"
+			>
+				<span className="navbar-toggler-icon"></span>
+			</button>
+			<div className="collapse navbar-collapse" id="navbarNav">
+				<ul className="navbar-nav">
+					<li className="nav-item">
+						<Link className="nav-link" to="/">
+							Home
+						</Link>
+					</li>
+					<li className="nav-item">
+						<Link className="nav-link" to="/automapper">
+							Automapper
+						</Link>
+					</li>
+					<li className="nav-item">
+						<Link className="nav-link" to="/fish_usa">
+							FishUSA
+						</Link>
+					</li>
+				</ul>
+			</div>
+		</div>
+	</nav>
+);
+
+export default Navbar;
